api/requirements.txt
Flask==2.3.2
Flask-Admin==1.6.1
Flask-BasicAuth==0.2.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
google-api-python-client==2.125.0
google-auth==2.29.0
psycopg2-binary==2.9.9
requests==2.31.0
gunicorn==21.2.0
python-dotenv==1.0.1
wtforms==3.1.2
wtforms-sqlalchemy==0.3
uvicorn[standard]==0.29.0
fastapi==0.110.2
httpx==0.27.0
pytz==2024.1
redis==5.0.4
openai>=1.0.0